Allow user to delete route
Closes #330

+++ rg2api.php
@@ -152,10 +152,10 @@ function handlePostRequest($type, $eventid) {
   $data = json_decode(file_get_contents('php://input'));
   $write = array();
   if (lockDatabase() !== FALSE) {
-    if ($type != 'addroute') {
+    if (($type != 'addroute') && ($type != 'deletemyroute')) {
       $loggedIn = logIn($data);
     } else {
-      // don't need to log in to add a route
+      // don't need to log in to add a route or delete your own
       $loggedIn = TRUE;
     }
     if ($loggedIn) {
@@ -194,11 +194,26 @@ function handlePostRequest($type, $eventid) {
        break;
 
       case 'deleteroute':
-        $write = deleteRoute($eventid);
+          // this is the manager delete function
+        $write = deleteMyRoute($eventid, $data);
         @unlink(CACHE_DIRECTORY."results_".$eventid.".json");
         @unlink(CACHE_DIRECTORY."tracks_".$eventid.".json");
         @unlink(CACHE_DIRECTORY."stats.json");
         break;
+
+        case 'deletemyroute':
+        // this is the user delete function
+        if (canDeleteMyRoute($eventid, $data)) {
+          $write = deleteRoute($eventid);
+          @unlink(CACHE_DIRECTORY."results_".$eventid.".json");
+          @unlink(CACHE_DIRECTORY."tracks_".$eventid.".json");
+          @unlink(CACHE_DIRECTORY."stats.json");            
+        } else {
+          $write["status_msg"] = "Delete failed";
+          $write["ok"] = FALSE;            
+        }
+        break;
+
 
       case 'deletecourse':
         $write = deleteCourse($eventid);
@@ -860,6 +875,8 @@ function deleteRoute($eventid) {
 
   if ($write["status_msg"] == "") {
     $write["ok"] = TRUE;
+    $write["eventid"] = $eventid;
+    $write["routeid"] = $routeid;
     $write["status_msg"] = "Route deleted";
     rg2log("Route deleted|".$eventid."|".$routeid);
   } else {
@@ -867,6 +884,26 @@ function deleteRoute($eventid) {
   }
 
   return($write);
+}
+
+function canDeleteMyRoute($eventid, $data) {
+  if (isset($_GET['routeid'])) {
+    $routeid = $_GET['routeid'];
+    $token = $data->token;
+    $filename = KARTAT_DIRECTORY."merkinnat_".$eventid.".txt";
+    $oldfile = file($filename);
+    $validrequest = false;
+    foreach ($oldfile as $row) {
+      $data = explode("|", $row);
+      if ($data[1] == $routeid) {
+        $hash = md5(serialize($row));
+        if ($hash == $token) {
+          $validrequest = true;
+        }
+      }
+    }
+  }
+  return $validrequest;
 }
 
 function addNewRoute($eventid, $data) {
@@ -916,6 +953,8 @@ function addNewRoute($eventid, $data) {
   }
 
   $newtrackdata = $data->courseid."|".$id."|".$name."|null|".$track."|".$controls.PHP_EOL;
+  // calculate a hash to allow deletion
+  $token = md5(serialize($newtrackdata));
 
   $newresultdata = "";
   if (($newresult === TRUE) || ($id >= GPS_RESULT_OFFSET)) {
@@ -1003,6 +1042,8 @@ function addNewRoute($eventid, $data) {
   if ($write["status_msg"] == "") {
     $write["ok"] = TRUE;
     $write["status_msg"] = "Record saved";
+    $write["token"] = $token;
+    $write["eventid"] = $eventid;
     //rg2log("Route saved|".$eventid."|".$id);
   } else {
     $write["ok"] = FALSE;
